+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 * Pages: hidden pages don't have the view-button anymore, as requested on http://forkcms.lighthouseapp.com/projects/61890/tickets/123 
 * Bugfix: Metanavigation subpages not shown in backend, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/129
 * Dashboard: Fixed issue with dashboard that wasn't scalling anymore, as reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/130
+* When FB-admin-ids are given, the facebook-statistics-tag is added on all pages
 
 2.0.2 (2010-11-24)
 --